app/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
auth0-ai==1.0.0b4
auth0-ai-langchain==1.0.0b4
auth0-fastapi==1.0.0b4
auth0-python==4.13.0
auth0-server-python==1.0.0b5
Authlib==1.6.5
black==25.9.0
blockbuster==1.5.25
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.14
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
ffmpy==0.6.3
filelock==3.20.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2025.9.0
googleapis-common-protos==1.71.0
gradio==5.49.1
gradio_client==1.13.3
grandalf==0.8
groovy==0.1.2
grpcio==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jwcrypto==1.5.6
langchain==1.0.0
langchain-core==1.0.0
langchain-openai==1.0.0
langchain-tavily==0.2.12
langchain-text-splitters==0.3.11
langgraph==1.0.0
langgraph-api==0.4.44
langgraph-checkpoint==2.1.2
langgraph-cli==0.4.4
langgraph-prebuilt==1.0.0
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.5.0
openfga-sdk==0.9.7
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
ruff==0.14.1
safehttpx==0.1.6
semantic-version==2.10.0
sentry-sdk==2.42.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
truststore==0.10.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
